[debug]
改一些小bug

+++ include/Key.h
@@ -139,9 +139,9 @@ inline Key int2key(uint64_t key) {
     a= a>>8;
     keylen++;
   }
-  for (int i = 1; i <= keylen; ++ i) {
+  for (int i = 1; i <=(int) keylen; ++ i) {
     auto shr = (keylen - i) * 8;
-    res.at(i - 1) = (shr >= 64u ? 0 : ((key >> shr) & ((1 << 8) - 1))); // Is equivalent to padding zero for short key
+    res.at(i - 1) = (shr >= 64u ? 0 : ((key >> (uint64_t)shr) & ((1 << 8) - 1))); // Is equivalent to padding zero for short key
   }
   res.at(64)=keylen;
   return res;
